Refactor out App component from Redux Provider
Redux has it's own was of passing everything
around, and best practice requires everything
to be wrapped in a 'Provider' Component.

Which works, because then App can be our
Container Component and live next to
all of the other components...or maybe it should
just be a Presentatinal component? Hmm....